// src/lib/types.ts
export type Role = 'ADMIN' | 'MANAGER' | 'MEMBER';
export type TaskType = 'TASK' | 'BUG' | 'STORY' | 'EPIC' | 'IMPROVEMENT';
export type Priority = 'LOWEST' | 'LOW' | 'MEDIUM' | 'HIGH' | 'HIGHEST';
export type ColumnCategory = 'TODO' | 'IN_PROGRESS' | 'DONE';